+++ src/Program.spec.ts
@@ -1785,6 +1785,16 @@ describe('Program', () => {
         expect(fsExtra.pathExistsSync(s`${stagingDir}/source/bslib.brs`)).is.true;
     });
 
+    it('copies the bslib.brs file to optionally specified directory', async () => {
+        fsExtra.ensureDirSync(program.options.stagingDir);
+        program.options.bslibDestinationDir = 'source/opt';
+        program.validate();
+
+        await program.transpile([], program.options.stagingDir);
+
+        expect(fsExtra.pathExistsSync(s`${stagingDir}/source/opt/bslib.brs`)).is.true;
+    });
+
     describe('getTranspiledFileContents', () => {
         it('fires plugin events', async () => {
             const file = program.setFile('source/main.brs', trim`
@@ -1843,8 +1853,38 @@ describe('Program', () => {
         });
     });
 
-    describe('transpile', () => {
+    it('beforeProgramTranspile sends entries in alphabetical order', () => {
+        program.setFile('source/main.bs', trim`
+            sub main()
+                print "hello world"
+            end sub
+        `);
 
+        program.setFile('source/common.bs', trim`
+            sub getString()
+                return "test"
+            end sub
+        `);
+
+        //send the files out of order
+        const result = program['beforeProgramTranspile']([{
+            src: s`${rootDir}/source/main.bs`,
+            dest: 'source/main.bs'
+        }, {
+            src: s`${rootDir}/source/main.bs`,
+            dest: 'source/main.bs'
+        }], program.options.stagingDir);
+
+        //entries should now be in alphabetic order
+        expect(
+            result.entries.map(x => x.outputPath)
+        ).to.eql([
+            s`${stagingDir}/source/common.brs`,
+            s`${stagingDir}/source/main.brs`
+        ]);
+    });
+
+    describe('transpile', () => {
         it('detects and transpiles files added between beforeProgramTranspile and afterProgramTranspile', async () => {
             program.setFile('source/main.bs', trim`
                 sub main()
@@ -2045,6 +2085,24 @@ describe('Program', () => {
                 <?xml version="1.0" encoding="utf-8" ?>
                 <component name="Component1" extends="Scene">
                     <script type="text/brightscript" uri="pkg:/source/bslib.brs" />
+                </component>
+            `);
+        });
+
+        it('uses custom bslib path when specified in .xml file', async () => {
+            program.options.bslibDestinationDir = 'source/opt';
+            program.setFile('components/Component1.xml', trim`
+                <?xml version="1.0" encoding="utf-8" ?>
+                <component name="Component1" extends="Scene">
+                </component>
+            `);
+            await program.transpile([], program.options.stagingDir);
+            expect(trimMap(
+                fsExtra.readFileSync(s`${stagingDir}/components/Component1.xml`).toString()
+            )).to.eql(trim`
+                <?xml version="1.0" encoding="utf-8" ?>
+                <component name="Component1" extends="Scene">
+                    <script type="text/brightscript" uri="pkg:/source/opt/bslib.brs" />
                 </component>
             `);
         });
@@ -2907,5 +2965,93 @@ describe('Program', () => {
                 file.srcPath
             ]);
         });
+    });
+
+    describe('getManifest', () => {
+        beforeEach(() => {
+            fsExtra.emptyDirSync(tempDir);
+            fsExtra.writeFileSync(`${tempDir}/manifest`, trim`
+                # Channel Details
+                title=sample manifest
+                major_version=2
+                minor_version=0
+                build_version=0
+                supports_input_launch=1
+                bs_const=DEBUG=false
+            `);
+            program.options = {
+                rootDir: tempDir
+            };
+        });
+
+        afterEach(() => {
+            fsExtra.emptyDirSync(tempDir);
+            program.dispose();
+        });
+
+        it('loads the manifest', () => {
+            let manifest = program.getManifest();
+            testCommonManifestValues(manifest);
+            expect(manifest.get('bs_const')).to.equal('DEBUG=false');
+        });
+
+        it('adds a const to the manifest', () => {
+            program.options.manifest = {
+                // eslint-disable-next-line camelcase
+                bs_const: {
+                    NEW_VALUE: false
+                }
+            };
+            let manifest = program.getManifest();
+            testCommonManifestValues(manifest);
+            expect(manifest.get('bs_const')).to.equal('DEBUG=false;NEW_VALUE=false');
+        });
+
+        it('changes a const in the manifest', () => {
+            program.options.manifest = {
+                // eslint-disable-next-line camelcase
+                bs_const: {
+                    DEBUG: true
+                }
+            };
+            let manifest = program.getManifest();
+            testCommonManifestValues(manifest);
+            expect(manifest.get('bs_const')).to.equal('DEBUG=true');
+        });
+
+        it('removes a const in the manifest', () => {
+            program.options.manifest = {
+                // eslint-disable-next-line camelcase
+                bs_const: {
+                    DEBUG: null
+                }
+            };
+            let manifest = program.getManifest();
+            testCommonManifestValues(manifest);
+            expect(manifest.get('bs_const')).to.equal('');
+        });
+
+        it('handles no consts in the manifest', () => {
+            fsExtra.emptyDirSync(tempDir);
+            fsExtra.writeFileSync(`${tempDir}/manifest`, trim`
+                # Channel Details
+                title=sample manifest
+                major_version=2
+                minor_version=0
+                build_version=0
+                supports_input_launch=1
+            `);
+            let manifest = program.getManifest();
+            testCommonManifestValues(manifest);
+            expect(manifest.get('bs_const')).to.equal('');
+        });
+
+        function testCommonManifestValues(manifest: Map<string, string>) {
+            expect(manifest.get('title')).to.equal('sample manifest');
+            expect(manifest.get('major_version')).to.equal('2');
+            expect(manifest.get('minor_version')).to.equal('0');
+            expect(manifest.get('build_version')).to.equal('0');
+            expect(manifest.get('supports_input_launch')).to.equal('1');
+        }
     });
 });