+++ src/util.ts
@@ -297,12 +297,17 @@ export class Util {
     public normalizeAndResolveConfig(config: BsConfig) {
         let result = this.normalizeConfig({});
 
-        //if no options were provided, try to find a bsconfig.json file
-        if (!config || !config.project) {
-            result.project = this.getConfigFilePath(config?.cwd);
+        if (config?.noProject) {
+            return result;
+        }
+
+        result.project = null;
+        if (config?.project) {
+            result.project = config?.project;
         } else {
-            //use the config's project link
-            result.project = config.project;
+            if (config?.cwd) {
+                result.project = this.getConfigFilePath(config?.cwd);
+            }
         }
         if (result.project) {
             let configFile = this.loadConfigFile(result.project, null, config?.cwd);
@@ -349,6 +354,11 @@ export class Util {
             config.logLevel = LogLevel[(config.logLevel as string).toLowerCase()];
         }
         config.logLevel = config.logLevel ?? LogLevel.log;
+        config.bslibDestinationDir = config.bslibDestinationDir ?? 'source';
+        if (config.bslibDestinationDir !== 'source') {
+            // strip leading and trailing slashes
+            config.bslibDestinationDir = config.bslibDestinationDir.replace(/^(\/*)(.*?)(\/*)$/, '$2');
+        }
         return config;
     }
 
@@ -504,6 +514,11 @@ export class Util {
      * ```
      */
     public rangesIntersect(a: Range, b: Range) {
+        //stop if the either range is misisng
+        if (!a || !b) {
+            return false;
+        }
+
         // Check if `a` is before `b`
         if (a.end.line < b.start.line || (a.end.line === b.start.line && a.end.character <= b.start.character)) {
             return false;
@@ -528,6 +543,10 @@ export class Util {
      * ```
      */
     public rangesIntersectOrTouch(a: Range, b: Range) {
+        //stop if the either range is misisng
+        if (!a || !b) {
+            return false;
+        }
         // Check if `a` is before `b`
         if (a.end.line < b.start.line || (a.end.line === b.start.line && a.end.character < b.start.character)) {
             return false;
@@ -551,6 +570,11 @@ export class Util {
     }
 
     public comparePositionToRange(position: Position, range: Range) {
+        //stop if the either range is misisng
+        if (!position || !range) {
+            return 0;
+        }
+
         if (position.line < range.start.line || (position.line === range.start.line && position.character < range.start.character)) {
             return -1;
         }
@@ -1209,9 +1233,9 @@ export class Util {
     /**
      * Copy the version of bslib from local node_modules to the staging folder
      */
-    public async copyBslibToStaging(stagingDir: string) {
+    public async copyBslibToStaging(stagingDir: string, bslibDestinationDir = 'source') {
         //copy bslib to the output directory
-        await fsExtra.ensureDir(standardizePath(`${stagingDir}/source`));
+        await fsExtra.ensureDir(standardizePath(`${stagingDir}/${bslibDestinationDir}`));
         // eslint-disable-next-line
         const bslib = require('@rokucommunity/bslib');
         let source = bslib.source as string;
@@ -1229,7 +1253,7 @@ export class Util {
             const position = positions[i];
             source = source.slice(0, position) + 'bslib_' + source.slice(position);
         }
-        await fsExtra.writeFile(`${stagingDir}/source/bslib.brs`, source);
+        await fsExtra.writeFile(`${stagingDir}/${bslibDestinationDir}/bslib.brs`, source);
     }
 
     /**